tests: mysql: rename utils file
